[change]content_attachment_validationの共通化
